fix: Change login flow
