--- src/client/containers/LoginContainer/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from "./useLogin";
